비밀번호 변경 기능 최초 등록 - 기존 로그인 시 사용되는 loginService 내 로그인 메소드 활용하여 기존 비밀번호 일치여부 확인 - 비밀번호 변경 시 솔트 재발급 및 비밀번호 등록 (기존 암호화 메소드 활용)
피드 삭제 시 로직 추가
- FileManger 내 파일 및 경로 삭제 메소드 추가
- 피드 삭제 시 실제 파일 및 경로 삭제, 좋아요 및 댓글 테이블 내 피드아이디 매칭되는 로우 삭제 메소드 추가
- 피드 삭제 시 기존 DB활용 실제 유저 여부 확인 -> 서비스에서 1차 유저 검증 후 DB검증 로직 추가

+++ src/main/java/com/routinergram/feed/service/FeedService.java
@@ -17,6 +17,7 @@ import com.routinergram.feed.repository.FeedRepository;
 import com.routinergram.interests.service.InterestsService;
 import com.routinergram.likes.service.LikesService;
 import com.routinergram.reply.domain.Reply;
+import com.routinergram.reply.repository.ReplyRepository;
 import com.routinergram.reply.service.ReplyService;
 import com.routinergram.user.service.UserNicknameService;
 
@@ -132,8 +133,25 @@ public class FeedService {
 	
 	public int deleteFeed(int UID, int FID) {
 		Feed feed = new Feed();
-		feed.setUID(UID);
+		
 		feed.setFID(FID);
-		return feedRepository.deleteFeed(feed);
+		int writerID = feedRepository.selectFeedByFID(feed).getUID();
+		if(writerID != UID) {
+			return 0;
+		}
+		feed.setUID(UID);
+		
+		String feedImage = feedRepository.selectFeedByFID(feed).getImage();
+		if(feedImage == null) {
+			replyService.deleteReplyByFID(FID);
+			likesService.deleteLikesByFID(FID);
+			return feedRepository.deleteFeed(feed);
+		} else {
+			if(FileManager.deleteFile(feedImage)) {
+				replyService.deleteReplyByFID(FID);
+				likesService.deleteLikesByFID(FID);
+				return feedRepository.deleteFeed(feed);
+			} else return 0;
+		}
 	}
 }
